Fix NPE
git-svn-id: file:///home/svn/DVA2006@6376 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/commands/cooler/NotificationListUpdateCommand.java
@@ -1,4 +1,4 @@
-// Copyright 2008 Global Virtual Airlines Group. All Rights Reserved.
+// Copyright 2008, 2009 Global Virtual Airlines Group. All Rights Reserved.
 package org.deltava.commands.cooler;
 
 import java.util.*;
@@ -6,17 +6,18 @@ import java.util.*;
 import org.deltava.commands.*;
 import org.deltava.dao.*;
 
+import org.deltava.util.CollectionUtils;
 import org.deltava.util.StringUtils;
 
 /**
  * A Web Site Command to update the Thread Notification list.
  * @author Luke
- * @version 2.1
+ * @version 2.7
  * @since 2.1
  */
 
 public class NotificationListUpdateCommand extends AbstractCommand {
-	
+
 	/**
 	 * Executes the command.
 	 * @param ctx the Command context
@@ -26,29 +27,31 @@ public class NotificationListUpdateCommand extends AbstractCommand {
 
 		// Get the threads to clear
 		Collection<String> IDs = ctx.getParameters("threadID");
-		try {
-			// Get the DAO and clear the notifications
-			SetCoolerNotification ndao = new SetCoolerNotification(ctx.getConnection());
-			ctx.startTX();
-			for (Iterator<String> i = IDs.iterator(); i.hasNext(); ) {
-				String id = i.next();
-				try {
-					int threadID = StringUtils.parseHex(id);
-					ndao.delete(threadID, ctx.getUser().getID());
-				} catch (NumberFormatException nfe) {
-					// empty
+		if (!CollectionUtils.isEmpty(IDs)) {
+			try {
+				// Get the DAO and clear the notifications
+				SetCoolerNotification ndao = new SetCoolerNotification(ctx.getConnection());
+				ctx.startTX();
+				for (Iterator<String> i = IDs.iterator(); i.hasNext();) {
+					String id = i.next();
+					try {
+						int threadID = StringUtils.parseHex(id);
+						ndao.delete(threadID, ctx.getUser().getID());
+					} catch (NumberFormatException nfe) {
+						// empty
+					}
 				}
+
+				// Commit
+				ctx.commitTX();
+			} catch (DAOException de) {
+				ctx.rollbackTX();
+				throw new CommandException(de);
+			} finally {
+				ctx.release();
 			}
-			
-			// Commit
-			ctx.commitTX();
-		} catch (DAOException de) {
-			ctx.rollbackTX();
-			throw new CommandException(de);
-		} finally {
-			ctx.release();
 		}
-		
+
 		// Forward to the JSP
 		CommandResult result = ctx.getResult();
 		result.setURL("notifythreads.do");
